implement jwt_claims, jwt_optional, etc, and format the code

+++ resources/item.py
@@ -1,5 +1,6 @@
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+
 from models.item import ItemModel
 
 
@@ -8,7 +9,7 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="Price cannot be blank.")
     parser.add_argument('store_id', type=int, required=True, help="Store is required.")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_item_by_name(name)
         if item:
@@ -27,7 +28,12 @@ class Item(Resource):
         #     return {'message': 'Error occurred when inserting a record'}, 500
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin is required'}, 401
+
         item = ItemModel.find_item_by_name(name)
         if item:
             item.delete_item_in_db()
@@ -47,6 +53,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
+        user_id = get_jwt_identity()
         items = ItemModel.get_items_from_db()
-        return {'items': [item.json() for item in items]}, 200
+        if user_id:
+            return {'items': [item.json() for item in items]}, 200
+        return {
+            'items': [item.name for item in items],
+            'message': 'more data will be available when you login'
+        }
